backend/api/requirements.txt
Django>=2.0, <3.0
djangorestframework>=3, <4.0
psycopg2>=2.7, <3.0
django-countries>5
django-cors-headers>3
django-tinymce>=2.8
Pillow>=7.0.0